Diagnostics tool
You have to make a Diagnostics object in your class that is testing
stuff. The methods to output data to the SmartDashboard are just the
names of the device you're testing. For example, to output a certain
servo's data, call the Servo method. You'll have to pass in the servo
and what you want it to be called on the dashboard. Every time you test
a new thing it makes a new label thing in the SmartDashboard. Let me
know if I have to add other methods to test other classes, right now
there's only like 4. Afterwords I'll edit the SmartDashboard on the
laptop we're using at competitions so it looks all nice, because when
you create a new label by calling methods it just puts it in the upper
left corner.